test(shell): split topbar-account-menu test file 1->7 to work around specter walker bug
specter v0.13.2's `coverage --strict` walker has a per-file-AC-credit
bug that prevented this PR from passing CI's spec-coverage gate. The
walker credits exactly one AC per source file even when:

  - The test file's runtime annotations (test names contain
    spec-id/AC-NN) all fire and pass.
  - specter ingest correctly writes 7 'passed' entries in
    .specter-results.json.
  - The source has 7 // @ac AC-NN annotations, structurally identical
    to other working multi-AC test files (e.g. host-detail-system-
    card.test.tsx, which credits 6/6 of its in-file ACs).

I exhausted the obvious workarounds:
  - renamed test file to match spec basename
  - stripped the AC traceability docstring header
  - renamed the spec file
  - flipped spec status to draft

None of them changed the 1/7 credited result. The walker behavior is
genuinely a specter quirk we can't address from the spec/test side.

Workaround: one AC per test file. specter's walker now sees exactly
one @ac per file and credits it correctly. Verified locally:

  $ specter ingest --go-test ... --junit ...
  $ specter coverage --strict | grep shell-account-menu
  frontend-shell-account-menu  T2  7  7  100%  PASS

Layout:
  shell-account-menu-helpers.ts        — TOPBAR_SRC snapshot, signIn,
                                         clearAuth (pure helpers only)
  shell-account-menu-ac01.test.tsx     — source-inspection
  shell-account-menu-ac02.test.tsx     — open + flip aria-expanded
  shell-account-menu-ac03.test.tsx     — Sign out menuitem present
  shell-account-menu-ac04.test.tsx     — Sign out fires logout + nav
  shell-account-menu-ac05.test.tsx     — Escape closes
  shell-account-menu-ac06.test.tsx     — click-outside closes
  shell-account-menu-ac07.test.tsx     — anon = no avatar

Each of the 6 rendering files declares its own vi.mock for
@tanstack/react-router + @/api/client because vitest's vi.mock hoister
only operates on the test file the call is in — moving them to the
helper would silently break the mocks (Cannot read properties of null
(reading 'navigate')).

Net behavior: identical test coverage. Same 7 assertions, same
mocks, same fixtures. No production code changed.

Original topbar-account-menu.test.tsx removed. The new file basenames
match the spec id, which also makes the directory listing self-
documenting (each AC -> one file).
