+++ app/frontend/tests/pages/shell-account-menu-ac02.test.tsx
@@ -1,0 +1,40 @@
+// @spec frontend-shell-account-menu
+// @ac AC-02
+
+import { describe, expect, test, beforeEach, vi } from 'vitest';
+import { fireEvent, render, screen } from '@testing-library/react';
+import { signIn, clearAuth } from './shell-account-menu-helpers';
+
+const navigateMock = vi.fn();
+vi.mock('@tanstack/react-router', async () => {
+  const actual = await vi.importActual<object>('@tanstack/react-router');
+  return {
+    ...actual,
+    useNavigate: () => navigateMock,
+    Link: ({ to, children, ...rest }: { to?: string; children: React.ReactNode } & Record<string, unknown>) =>
+      ({ type: 'a', props: { href: to, ...rest, children }, key: null, ref: null } as any),
+  };
+});
+vi.mock('@/api/client', () => ({
+  default: { POST: vi.fn(async () => ({ response: { ok: true, status: 204 } })) },
+}));
+
+import { TopBar } from '@/components/shell/TopBar';
+
+beforeEach(() => {
+  clearAuth();
+  navigateMock.mockReset();
+});
+
+describe('frontend-shell-account-menu — behavioral', () => {
+  test('frontend-shell-account-menu/AC-02 — clicking the avatar opens the menu and flips aria-expanded', () => {
+    signIn();
+    render(<TopBar />);
+    const btn = screen.getByRole('button', { name: /account/i });
+    expect(btn).toHaveAttribute('aria-expanded', 'false');
+    expect(screen.queryByRole('menu')).toBeNull();
+    fireEvent.click(btn);
+    expect(btn).toHaveAttribute('aria-expanded', 'true');
+    expect(screen.getByRole('menu')).toBeInTheDocument();
+  });
+});
