+++ app/frontend/tests/pages/shell-account-menu-ac07.test.tsx
@@ -1,0 +1,35 @@
+// @spec frontend-shell-account-menu
+// @ac AC-07
+
+import { describe, expect, test, beforeEach, vi } from 'vitest';
+import { render, screen } from '@testing-library/react';
+import { clearAuth } from './shell-account-menu-helpers';
+
+const navigateMock = vi.fn();
+vi.mock('@tanstack/react-router', async () => {
+  const actual = await vi.importActual<object>('@tanstack/react-router');
+  return {
+    ...actual,
+    useNavigate: () => navigateMock,
+    Link: ({ to, children, ...rest }: { to?: string; children: React.ReactNode } & Record<string, unknown>) =>
+      ({ type: 'a', props: { href: to, ...rest, children }, key: null, ref: null } as any),
+  };
+});
+vi.mock('@/api/client', () => ({
+  default: { POST: vi.fn(async () => ({ response: { ok: true, status: 204 } })) },
+}));
+
+import { TopBar } from '@/components/shell/TopBar';
+
+beforeEach(() => {
+  clearAuth();
+  navigateMock.mockReset();
+});
+
+describe('frontend-shell-account-menu — behavioral', () => {
+  test('frontend-shell-account-menu/AC-07 — without identity the avatar button does not render', () => {
+    // useAuthStore identity is null by default after clearAuth.
+    render(<TopBar />);
+    expect(screen.queryByRole('button', { name: /account/i })).toBeNull();
+  });
+});
